Reimplement GetUserOrgsList to make it simple and clear
